feat: 移除 AR 引导相关视图，整合视图逻辑
